requirements.txt
sentence-transformers
faiss-cpu
langchain
Flask
PyPDF2
pdfplumber
rq
websockets
openai